+++ src/factories/document.ts
@@ -10,8 +10,8 @@ import { createTrailingCommentAttachable } from "./trailing-comment-attachable.j
 
 export function createDocument(
   position: Position,
-  startMark: boolean,
-  endMark: boolean,
+  directivesEndMarker: boolean,
+  documentEndMarker: boolean,
   head: DocumentHead,
   body: DocumentBody,
   trailingComment: null | Comment,
@@ -19,8 +19,8 @@ export function createDocument(
   return {
     ...createNode("document", position),
     ...createTrailingCommentAttachable(trailingComment),
-    startMark,
-    endMark,
+    directivesEndMarker,
+    documentEndMarker,
     children: [head, body],
   };
 }