+++ src/types.ts
@@ -119,8 +119,8 @@ export interface Root extends Parent {
 
 export interface Document extends Parent, TrailingCommentAttachable {
   type: "document";
-  startMark: boolean;
-  endMark: boolean;
+  directivesEndMarker: boolean;
+  documentEndMarker: boolean;
   children: [DocumentHead, DocumentBody];
 }